requirements.txt
altair
numpy
pandas
Pillow
plotly
scipy
scikit-learn>=0.18
streamlit
yfinance
